refactor: Split is valid guild to other file
Prevent circular import when using in cog or other file
